Python Dev/WEEK 3-4/Weather Predictor/requirements.txt
requests>=2.31.0
pandas>=2.2.2
matplotlib>=3.8.4
scikit-learn>=1.3.0
python-dotenv>=1.0.1
